@@ -28,30 +28,31 @@
     "azure storage table"
   ],
   "dependencies": {
-    "@azure/data-tables": "^1.0.0-beta.2",
     "azure-storage": "^2.10.3"
   },
   "devDependencies": {
+    "@azure/data-tables": "^1.0.0-beta.3",
     "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@types/chai": "^4.2.12",
-    "@types/mocha": "^8.0.2",
-    "@types/node": "^14.0.27",
-    "@typescript-eslint/eslint-plugin": "^3.9.0",
-    "azurite": "^2.7.1",
+    "@types/chai": "^4.2.14",
+    "@types/mocha": "^8.0.4",
+    "@types/node": "^14.14.10",
+    "@typescript-eslint/eslint-plugin": "^4.9.0",
+    "azurite": "^3.9.0",
     "chai": "^4.2.0",
-    "eslint": "^7.7.0",
-    "eslint-config-standard-with-typescript": "^18.0.2",
-    "eslint-plugin-import": "^2.22.0",
+    "eslint": "^7.14.0",
+    "eslint-config-standard-with-typescript": "^19.0.1",
+    "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
+    "eslint-plugin-standard": "^4.1.0",
     "gulp": "^4.0.2",
     "gulp-shell": "^0.8.0",
     "local-web-server": "^4.2.1",
+    "managed-service-daemon": "^1.0.1",
     "minimist": "^1.2.5",
-    "mocha": "^8.1.1",
+    "mocha": "^8.2.1",
     "nyc": "^15.1.0",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7"
+    "ts-node": "^9.1.0",
+    "typescript": "^4.1.2"
   }
 }
